+++ src/main/kotlin/library/self/ForbiddenFour.kt
@@ -1,34 +1,23 @@
-package domain.library
+package domain.library.self
 
 import domain.CoordinateState
 import domain.Position
 
-object ExceedFive {
-    fun isExceedFive(
-        board: List<List<CoordinateState>>,
-        position: Position,
-        coordinateState: CoordinateState,
-    ): Boolean {
+object ForbiddenFour {
+    fun isForbiddenFour(board: List<List<CoordinateState>>, position: Position): Boolean {
         var fourStone = 0
-        fourStone += fourORjang1(board, position.x, position.y, coordinateState)
-        fourStone += fourORjang2(board, position.x, position.y, coordinateState)
-        fourStone += fourORjang3(board, position.x, position.y, coordinateState)
-        fourStone += fourORjang4(board, position.x, position.y, coordinateState)
-        return fourStone >= 1
+        fourStone += fourORjang1(board, position.x, position.y)
+        fourStone += fourORjang2(board, position.x, position.y)
+        fourStone += fourORjang3(board, position.x, position.y)
+        fourStone += fourORjang4(board, position.x, position.y)
+        return fourStone >= 2
     }
 
-    private fun fourORjang1(
-        board: List<List<CoordinateState>>,
-        x: Int,
-        y: Int,
-        targetCoordinateState: CoordinateState,
-    ): Int {
+    private fun fourORjang1(board: List<List<CoordinateState>>, x: Int, y: Int): Int {
         var stone1 = 0
         var stone2 = 0
         var allStone = 0
         var blink1 = 1
-        val nonTargetCoordinateState =
-            if (targetCoordinateState == CoordinateState.BLACK) CoordinateState.WHITE else CoordinateState.BLACK
 
         // ←  탐색
         var yy = y
@@ -36,11 +25,11 @@ object ExceedFive {
         var check = false
         while (true) {
             if (xx == -1) break
-            if (board[yy][xx] == targetCoordinateState) {
+            if (board[yy][xx] == CoordinateState.BLACK) {
                 check = false
                 stone1++
             }
-            if (board[yy][xx] == nonTargetCoordinateState) break
+            if (board[yy][xx] == CoordinateState.WHITE) break
             if (board[yy][xx] == CoordinateState.EMPTY) {
                 check = if (check == false) {
                     true
@@ -64,11 +53,11 @@ object ExceedFive {
         check = false
         while (true) {
             if (xx == 15) break
-            if (board[yy][xx] == targetCoordinateState) {
+            if (board[yy][xx] == CoordinateState.BLACK) {
                 check = false
                 stone2++
             }
-            if (board[yy][xx] == nonTargetCoordinateState) break
+            if (board[yy][xx] == CoordinateState.WHITE) break
             if (board[yy][xx] == CoordinateState.EMPTY) {
                 check = if (check == false) {
                     true
@@ -88,23 +77,19 @@ object ExceedFive {
 
         // 사사찾는 트리거
 
-        // 현재놓은돌 +1 +5 => 6목이상은 장목. 여기서 놓은돌기준 두방향모두 돌이있어야 장목
-        return if (allStone >= 5 && stone1 != 0 && stone2 != 0) 1 else 0
+        return if (allStone != 3) {
+            0 // 놓은돌제외 3개아니면 4가아니니까.
+        } else {
+            1 // 놓은돌제외 3개면 4임. 닫히고 열린지는 상관없음.
+        }
     }
 
     // ↖ ↘ 탐색
-    private fun fourORjang2(
-        board: List<List<CoordinateState>>,
-        x: Int,
-        y: Int,
-        targetCoordinateState: CoordinateState,
-    ): Int {
+    private fun fourORjang2(board: List<List<CoordinateState>>, x: Int, y: Int): Int {
         var stone1 = 0
         var stone2 = 0
         var allStone = 0
         var blink1 = 1
-        val nonTargetCoordinateState =
-            if (targetCoordinateState == CoordinateState.BLACK) CoordinateState.WHITE else CoordinateState.BLACK
 
         // ↖  탐색
         var yy = y - 1
@@ -112,11 +97,11 @@ object ExceedFive {
         var check = false
         while (true) {
             if (xx == -1 || yy == -1) break
-            if (board[yy][xx] == targetCoordinateState) {
+            if (board[yy][xx] == CoordinateState.BLACK) {
                 check = false
                 stone1++
             }
-            if (board[yy][xx] == nonTargetCoordinateState) break
+            if (board[yy][xx] == CoordinateState.WHITE) break
             if (board[yy][xx] == CoordinateState.EMPTY) {
                 check = if (check == false) {
                     true
@@ -141,11 +126,11 @@ object ExceedFive {
         var blink2 = blink1
         while (true) {
             if (xx == 15 || yy == 15) break
-            if (board[yy][xx] == targetCoordinateState) {
+            if (board[yy][xx] == CoordinateState.BLACK) {
                 check = false
                 stone2++
             }
-            if (board[yy][xx] == nonTargetCoordinateState) break
+            if (board[yy][xx] == CoordinateState.WHITE) break
             if (board[yy][xx] == CoordinateState.EMPTY) {
                 check = if (check == false) {
                     true
@@ -164,22 +149,15 @@ object ExceedFive {
         }
         allStone = stone1 + stone2
 
-        return if (allStone >= 5 && stone1 != 0 && stone2 != 0) 1 else 0
+        return if (allStone != 3) 0 else 1
     }
 
     // ↑ ↓ 탐색
-    private fun fourORjang3(
-        board: List<List<CoordinateState>>,
-        x: Int,
-        y: Int,
-        targetCoordinateState: CoordinateState,
-    ): Int {
+    private fun fourORjang3(board: List<List<CoordinateState>>, x: Int, y: Int): Int {
         var stone1 = 0
         var stone2 = 0
         var allStone = 0
         var blink1 = 1
-        val nonTargetCoordinateState =
-            if (targetCoordinateState == CoordinateState.BLACK) CoordinateState.WHITE else CoordinateState.BLACK
 
         // ↑  탐색
         var yy = y - 1
@@ -187,11 +165,11 @@ object ExceedFive {
         var check = false
         while (true) {
             if (yy == -1) break
-            if (board[yy][xx] == targetCoordinateState) {
+            if (board[yy][xx] == CoordinateState.BLACK) {
                 check = false
                 stone1++
             }
-            if (board[yy][xx] == nonTargetCoordinateState) break
+            if (board[yy][xx] == CoordinateState.WHITE) break
             if (board[yy][xx] == CoordinateState.EMPTY) {
                 check = if (check == false) {
                     true
@@ -215,11 +193,11 @@ object ExceedFive {
         var blink2 = blink1
         while (true) {
             if (yy == 15) break
-            if (board[yy][xx] == targetCoordinateState) {
+            if (board[yy][xx] == CoordinateState.BLACK) {
                 check = false
                 stone2++
             }
-            if (board[yy][xx] == nonTargetCoordinateState) break
+            if (board[yy][xx] == CoordinateState.WHITE) break
             if (board[yy][xx] == CoordinateState.EMPTY) {
                 check = if (check == false) {
                     true
@@ -237,32 +215,25 @@ object ExceedFive {
         }
         allStone = stone1 + stone2
 
-        return if (allStone >= 5 && stone1 != 0 && stone2 != 0) 1 else 0
+        return if (allStone != 3) 0 else 1
     }
 
-    private fun fourORjang4(
-        board: List<List<CoordinateState>>,
-        x: Int,
-        y: Int,
-        targetCoordinateState: CoordinateState,
-    ): Int {
+    private fun fourORjang4(board: List<List<CoordinateState>>, x: Int, y: Int): Int {
         var stone1 = 0
         var stone2 = 0
         var allStone = 0
         var blink1 = 1
-        val nonTargetCoordinateState =
-            if (targetCoordinateState == CoordinateState.BLACK) CoordinateState.WHITE else CoordinateState.BLACK
 
         var yy = y - 1
         var xx = x + 1
         var check = false
         while (true) {
             if (xx == 15 || yy == -1) break
-            if (board[yy][xx] == targetCoordinateState) {
+            if (board[yy][xx] == CoordinateState.BLACK) {
                 check = false
                 stone1++
             }
-            if (board[yy][xx] == nonTargetCoordinateState) break
+            if (board[yy][xx] == CoordinateState.WHITE) break
             if (board[yy][xx] == CoordinateState.EMPTY) {
                 check = if (check == false) {
                     true
@@ -287,11 +258,11 @@ object ExceedFive {
         var blink2 = blink1
         while (true) {
             if (xx == -1 || yy == 15) break
-            if (board[yy][xx] == targetCoordinateState) {
+            if (board[yy][xx] == CoordinateState.BLACK) {
                 check = false
                 stone2++
             }
-            if (board[yy][xx] == nonTargetCoordinateState) break
+            if (board[yy][xx] == CoordinateState.WHITE) break
             if (board[yy][xx] == CoordinateState.EMPTY) {
                 check = if (check == false) {
                     true
@@ -310,6 +281,6 @@ object ExceedFive {
         }
         allStone = stone1 + stone2
 
-        return if (allStone >= 5 && stone1 != 0 && stone2 != 0) 1 else 0
+        return if (allStone != 3) 0 else 1
     }
 }
